Add Docker support and update project settings
Updated `FavoritesMovieBackend.csproj` to include `UserSecretsId` and `DockerDefaultTargetOS` properties, and added a package reference for `Microsoft.VisualStudio.Azure.Containers.Tools.Targets`.

Modified `launchSettings.json` to remove old schema and IIS settings, and added new profiles for HTTP, HTTPS, and Docker with environment variables and URLs.

Created a new `.dockerignore` file to specify files and directories to be ignored by Docker, including configuration files, `bin`, `obj`, and other development-related files.

Added a new `Dockerfile` to define the Docker image for the application, including steps to build and publish the application using the .NET 8.0 SDK and runtime images, and set up the container to expose ports 8080 and 8081.